==> 01_precip_sta_map.Rmd <==
---
title: "Untitled"
author: "CJ Tinant"
date: "10/23/2019"
output: html_document
---

# purpose 
Prepare map figures for dissertation  chapters.
# codebook  
<!--  
weather station and stream gage data====   
sta           NOAA weather station locations  
  _meta       metadata  
  _fin        final stations used in the analysis  
  _orig       all stations (n = 46 sations)  
  _plus       potential stations (n = 14 stations)  
wsd  
  _summary    summary watershed locations-- includes gaged & ungaged wsds  
gage  
  _full       monthly unit streamflow data with complete records (wy 1989-2017) 
  
mapping variables==== 
voronoi       Theissen polygons for stations  
usa           lower-48 states  
states        lower-48 states & counties 
counties      lower-48 counties  
  _sd         South Dakota  
  _ne         Nebraska  
  _wy         Wyoming  
  _bound      boundary -- a state line  
bbox          bounding box  
usmap         USA map as a grob  
-->  

```{r setup_&_library, message=FALSE}  
knitr::opts_chunk$set(echo = FALSE)   
options(tibble.print_max = 70) # sets tibble output for printing       

# Set up the library of packages----     
library("conflicted")        # An alternative conflict resolution strategy  
library("here")              # identifies where to save work  
library("rio")               # more robust I/O - to import and clean data  
library('deldir')            # for Vorononi tesselation - Theissen polygons  
library("cowplot")           # features for creating publication-quality figs  
library("tidyverse")    

# resolve conflicts among package====  
conflict_prefer("filter", "dplyr")  
conflict_prefer("filter", "dplyr")  

``` 

```{r get_data}  

# Import weather station data----    
sta_meta_orig <- import("data/sta_meta_orig.csv")   
sta_meta_plus <- import("data/sta_meta_plus.csv")   
sta_meta_fin  <- import("data/sta_meta_fin.csv")  

# Import stream gage data====  
wsd_summary <- import("data/wsd_summary.csv") #%>%  
#  rename(sta = sta_id)  

gage_full <- import("data/gage_mon_full.csv")  

# identify complete stations for plotting====  
gage_full <- semi_join(wsd_summary, gage_full,  
                            by = "sta") %>%  
  mutate(complete = "yes") %>%  
  mutate(type = "gaged") %>%  
  select(sta, type, complete)  

wsd_summary <- full_join(wsd_summary, gage_full,  
                          by = c("type", "sta") 
                          ) %>%  
  select(sta,  
         type,  
         ecoreg,  
         complete,  
         dec_lat,  
         dec_lon  
         ) %>%   
  mutate(type2 = case_when(  
    complete   == "yes" ~ "complete",  
    type       == "gaged" ~ "incomplete",  
    TRUE ~ type  
    ))  %>%  
  select(-c(complete, type)) %>%  
  rename(type = type2)  

rm(gage_full)  

```

```{r prepare_plot_data, eval=FALSE}

# Prepare map elements----  
# Create Theissen line segments====      
voronoi        <- deldir(sta_meta_fin$lon, sta_meta_fin$lat)    

# Order stations for plotting====   
sta_meta_orig <- sta_meta_orig %>%  
  mutate(group = fct_relevel(group,   
                             "NW",  
                             "NC",  
                             "NE",  
                             "SW",  
                             "SC",  
                             "SE" ))   

sta_meta_plus <- sta_meta_plus %>%  
  mutate(group = fct_relevel(group,   
                             "NW",  
                             "NC",  
                             "NE",  
                             "SW",  
#                             "SC",  # not sure why no SC  
                             "SE"))   

sta_meta_fin <- sta_meta_fin %>%  
  mutate(group = fct_relevel(group,   
                             "NW",  
                             "NC",  
                             "NE",  
                             "SW",  
                             "SC",  
                             "SE"))   

# prepare full gages for plotting  
gage_full_pl <- wsd_summary %>%  
  filter(type == "complete")  

# Define the study area using data from the 'maps' package-----   
# Import polygon data for maps====   
usa            <- map_data("usa")    
states         <- map_data("state")   
counties       <- map_data("county")   

# Filter counties for mapping -- subset SD counties ====  
#   need to filter counties for SD & NE separately because of duplicate vals   

counties_sd <- subset(counties,  
                      region %in%   
                        c("south dakota"))  
     
counties_sd <- subset(counties_sd,  
                      subregion %in%  
                        c("butte",  
                          "meade",  
                          "lawrence",  
                          "pennington",  
                          "custer",  
                          "fall river",  
                          "shannon",  
                          "jackson",  
                          "bennett",  
                          "haakon",    
                          "todd",     
                          "mellette",  
                          "jones",  
                          "stanley",  
                          "lyman",  
                          "tripp",  
                          "hughes",  
                          "ziebach",  
                          "dewey",  
                          "hughes",  
                          "sully",   
                          "potter"))    

#   subset NE counties####  
counties_ne <- subset(counties,   
                      region %in%  
                        c("nebraska"))  

counties_ne <- subset(counties_ne,   
                      subregion %in%  
                        c("sioux",   
                          "dawes",   
                          "sheridan",   
                          "box butte",    
                          "cherry",  
                          "brown",   
                          "keya paha",    
                          "rock"))   

#   join SD & NE counties####   
counties <- bind_rows(counties_sd, counties_ne)  
rm(counties_sd, counties_ne)   

# Create SD-NE & SD-WY lines====   
ne_bound <- subset(counties,   
                   subregion %in%     
                     c("fall river",   
                       "shannon",  
                       "bennett",   
                       "todd",   
                       "tripp")) %>%  
  filter(lat < 43.1)      

wy_bound <- subset(counties,   
                   subregion %in%  
                     c("butte",      
                       "lawrence"  ,  
                       "pennington" ,  
                       "custer",      
                       "fall river",  
                       "sioux"  
                     )) %>%  
  filter(long < -104)  

# Create reservation counties====  
rez <- subset(counties,  
              subregion %in%  
                c("shannon",      
                 "bennett",  
                 "jackson",  
                 "mellette",  
                 "todd",  
                 "dewey",  
                 "ziebach"))  

ridge_bound <- subset(counties,   
                      subregion %in%  
                        c("shannon",      
                          "bennett",  
                          "jackson"))  

rosebud_bound <- subset(counties,   
                   subregion %in%  
                     c("mellette",  
                       "todd"))  


# Prepare ecoregions for plotting====  
gages <- wsd_summary %>%  
  filter(ecoreg != "") %>%  
    mutate(ecoreg = as.factor(ecoreg)) %>%     
    mutate(ecoreg = fct_relevel(ecoreg,   
                             "Pierre Shale Plains",  
                             "Pine Ridge Escarpment",  
                             "White River Badlands",    
                             "Black Hills Plateau",   
                             "Keya Paha Tablelands",   
                             "Sand Hills"))  

```

```{r prepare_USA-grob}

# create a bounding box -- order is SW, NW, NE, SE====    
bbox <- tibble(lat = c(42.0, 45.0, 45.0, 42.0),  
               long = c(-104.5, -104.5, -99.5, -99.5),  
               group = c(1, 1, 1, 1))  

# plot US map====     
usmap <- ggplot() +  
  geom_polygon(data = usa,  
               aes(x=long, y = lat, group = group),   
               fill = "gray70",  
               color = "gray70") +       
  coord_fixed(1.3) +  
  theme_nothing() +  
  geom_polygon(data = bbox,  
               aes(x=long, y = lat, group = group),  
               fill = "gray30",  
               color = "gray30") +   
  xlab("") +   
  ylab("")   

# change the US map into a grob####   
usmap <- as_grob(usmap, device = NULL)  
  
rm(usa)  

```

```{r prepare_data_duplicated-above, eval=FALSE}
 
# import stream gage data====  
wsd_summary <- import("data/wsd_summary.csv") %>% 
  rename(sta = sta_id) 
 
gage_full <- import("data/gage_mon_full.csv")  
 
# identify complete stations for plotting====  
gage_full <- semi_join(wsd_summary, gage_full,  
                            by = "sta") %>%  
  mutate(complete = "yes") %>%  
  mutate(type = "gaged") %>%  
  select(sta, type, complete)  
 
wsd_summary <- full_join(wsd_summary, gage_full, 
                          by = c("type", "sta") 
                          ) %>% 
  select(sta,  
         type,  
         complete,  
         dec_lat,  
         dec_lon, 
         ecoreg 
         ) %>%   
  mutate(type2 = case_when(  
    complete   == "yes" ~ "complete",  
    type       == "gaged" ~ "incomplete", 
    TRUE ~ type 
    ))  %>% 
  select(-c(complete, type)) %>% 
  rename(type = type2) %>% 
  mutate(ecoreg = case_when(  
    ecoreg == "" ~ "ungaged",  
    TRUE ~ ecoreg)  
    ) %>% 
  # remove ungaged stations 
  filter(type != "ungaged") 
 
rm(gage_full) 
 
# prepare map elements----  
 
# Define the study area using data from the 'maps' package-----   
# import polygon data for maps====   
usa            <- map_data("usa")    
states         <- map_data("state")   
counties       <- map_data("county")   
   
# filter counties for mapping====  
#   need to filter counties for SD & NE separately because of duplicate vals   
#   subset SD counties####   
counties_sd <- subset(counties,  
                      region %in%   
                        c("south dakota"))  
     
counties_sd <- subset(counties_sd,  
                      subregion %in%  
                        c("butte",  
                          "meade",  
                          "lawrence",  
                          "pennington",  
                          "custer",  
                          "fall river",  
                          "shannon",  
                          "jackson",  
                          "bennett",  
                          "haakon",    
                          "todd",     
                          "mellette",  
                          "jones",  
                          "stanley",  
                          "lyman",  
                          "tripp",  
                          "hughes",  
                          "ziebach",  
                          "dewey",  
                          "hughes",  
                          "sully",   
                          "potter"     
                        )  
)    
   
#   subset NE counties####  
counties_ne <- subset(counties,   
                      region %in%  
                        c("nebraska"))  
   
counties_ne <- subset(counties_ne,   
                      subregion %in%  
                        c("sioux",   
                          "dawes",   
                          "sheridan",   
                          "box butte",    
                          "cherry",  
                          "brown",   
                          "keya paha",    
                          "rock"        
                        )   
)   

#   join SD & NE counties####   
counties <- bind_rows(counties_sd, counties_ne)  
rm(counties_sd, counties_ne)   
 
# create SD-NE & SD-WY lines====   
ne_bound <- subset(counties,   
                   subregion %in%     
                     c("fall river",   
                       "shannon",  
                       "bennett",   
                       "todd",   
                       "tripp")) %>%  
  filter(lat < 43.1)      
  
wy_bound <- subset(counties,   
                   subregion %in%  
                     c("butte",      
                       "lawrence"  ,  
                       "pennington" ,  
                       "custer",      
                       "fall river",  
                       "sioux"  
                     )) %>%  
  filter(long < -104)  
     
# create Reservation counties==== 
rez <- subset(counties,   
                   subregion %in%  
                     c("shannon",      
                       "bennett"  ,  
                       "jackson", 
                       "mellette", 
                       "todd", 
                       "dewey", 
                       "ziebach" 
                     )) 
 
ridge_bound <- subset(counties,   
                   subregion %in%  
                     c("shannon",      
                       "bennett"  ,  
                       "jackson") 
                   )  

rosebud_bound <- subset(counties,   
                   subregion %in%  
                     c("mellette", 
                       "todd") 
                       ) 

# plot US map as a grob----   
# create a bounding box -- order is SW, NW, NE, SE====    
bbox <- tibble(lat = c(42.0, 45.0, 45.0, 42.0),  
               long = c(-104.5, -104.5, -99.5, -99.5),  
               group = c(1, 1, 1, 1)  
)  
  
# plot US map====     
usmap <- ggplot() +  
  geom_polygon(data = usa,  
               aes(x=long, y = lat, group = group),   
               fill = "gray70",  
               color = "gray70"  
  ) +       
  coord_fixed(1.3) +  
  theme_nothing() +  
  geom_polygon(data = bbox,  
               aes(x=long, y = lat, group = group),  
               fill = "gray30",  
               color = "gray30") +   
  xlab("") +   
  ylab("")   
     
# change the US map into a grob####   
usmap <- as_grob(usmap, device = NULL)  
  
rm(usa)  
  
```

```{r plot_theissen-polygons_weather_station_map, eval=FALSE}

# plot weather station map====    
# Legend.position can be a numeric vector c(x,y), where x and y are the   
#   coordinates of the legend box between 0 and 1.  
#     c(0,0) corresponds to the “bottom left” and   
#     c(1,1) corresponds to the “top right” position.  

ggplot() +   
  theme_classic() +  
  coord_fixed(1.3) +  
  xlab("") +    
  ylab("") +      
  theme(axis.line         = element_line(colour = "grey60")) +  
  theme(axis.ticks        = element_line(colour = "grey60")) +  
  theme(axis.text         = element_text(colour = "grey50")) +  
  theme(legend.position   = c(1.1, 0.4)) +     
  theme(legend.title      = element_text(size = 10)) +   
  theme(legend.text       = element_text(size = 8)) +  
  theme(legend.key.height =  unit(0.5, 'cm')) +  
  guides(fill  = guide_legend(order = 2),  
         color = guide_legend(order = 1)) +     
  # add counties      
  geom_polygon(data = counties,  
               aes(x = long, y = lat,     
                   group = group),   
               color     = "gray80",  
               fill      = "NA") +   
  # add NE stateline boundary    
  geom_line(data = ne_bound,  
            aes(x = long, y = lat,     
                group = group),  
            color = "gray60") +   
  #   # add WY stateline boundary    
  geom_line(data = wy_bound,   
            aes(x = long, y = lat,    
                group = group),  
            color = "gray60") +    
  # add stream gaging stations     
geom_point(data = gage_full_pl,  
             aes(dec_lon, dec_lat,   
                 fill = factor(gage_full_pl$type)),   
             shape    = 17,  
             size     = 1.5,    
             stroke   = 0.5,  
             color    = "gray50") +      
  
# edited out the short-record gaging stations====  
#  geom_point(data = wsd_summary,  
#             aes(dec_lon, dec_lat,   
#                 fill = factor(wsd_summary$type)),   
 #            shape    = 24,  
 #            size     = 0.8,    
#             stroke   = 0.5,  
#             color    = "gray50") +      
  scale_fill_grey(start = 1.0,    
                  end   = 1.0,    
                  na.value   = "red",   
                  aesthetics = "fill",    
                  name       = 'Streamflow records',   
                  guide      = 'legend') +   
  # add orig stations   
  geom_point(data = sta_meta_orig,   
             aes(x = longitude, y = latitude,   
                 color = factor(sta_meta_orig$group)),    
             size = 1.5) +       
  # add plus stations    
  geom_point(data = sta_meta_plus,  
             aes(x = longitude, y = latitude,   
                 color  = factor(sta_meta_plus$group)),       
             size = 2) +  
  scale_colour_grey(start = 0.0,    
                    end   = 0.8,    
                    na.value   = "red",   
                    aesthetics = "colour",  
                    name       = 'Station groups',  
                    guide      = 'legend') +   
  # add final stations     
  geom_point(data = sta_meta_fin,   
             aes(x = longitude, y = latitude,   
                 color = factor(sta_meta_fin$group)),       
             size = 3) +   
  # add Theissen polygons   
  geom_segment(data = voronoi$dirsgs,   
               aes(x    = x1, y    = y1,   
                   xend = x2, yend = y2),   
               size     = 1,    
               linetype = 1,    
               color    = "gray35") +   
  # add labels to stations  
  geom_text(data = sta_meta_fin,    
            aes(x = longitude, y = latitude,  
                label = group),    
            vjust         = -0.8,  
            check_overlap = TRUE) +      
  # add the US map as a grob    
  annotation_custom(grob = usmap,    
                    xmin = -99.3, xmax = -97.6,   
                    ymin =  44.5, ymax = 45.5)    

#rm(bbox, counties,   
#   sta_meta_fin,  
#   wsd_summary,  
#   ne_bound,  
#   sta_meta_plus,   
#   states,   
#   usmap,   
#   voronoi,   
#   wy_bound,  
#   sta_meta_orig)    

```

```{r plot-map}
# export map & clean up   
ggplot2::ggsave(filename = "figure/sta-map.png",    
                width    = 7.5,  
                height   = 4.5,  
                units    = "in")     
```

```{r plot_stream-gages-map}
# plot  station map====    
# Legend.position can be a numeric vector c(x,y), where x and y are the   
#   coordinates of the legend box between 0 and 1.  
#     c(0,0) corresponds to the “bottom left” and   
#     c(1,1) corresponds to the “top right” position.  
# remember 'trouble' -- 't', 'r', 'b', 'l'  

ggplot() +   
  theme_classic() +  
  coord_fixed(1.3) +  
  xlab("") +    
  ylab("") +   
  theme(axis.line  = element_line(colour = "grey60")) +  
  theme(axis.ticks = element_line(colour = "grey60")) +  
  theme(axis.text  = element_text(colour = "grey50")) +  
  theme(plot.margin=grid::unit(c(0,20,0,-20), "mm")) +
  # create legend elements====  
  scale_colour_grey(name    = 'Ecoregion type',  
                    guide   = 'legend') +   
  scale_shape_discrete(name  = 'Gage completeness',   
                       guide = 'legend') +   
  # add the US map as a grob    
  annotation_custom(grob = usmap,    
                    xmin = -98.9,    
                    xmax = -96.9,   
                    ymin = 44.5,   
                    ymax = 45.5) +   
  theme(legend.position   = c(1.2, 0.4)) +    
  theme(legend.title      = element_text(size = 10)) +   
  theme(legend.text       = element_text(size = 8)) +  
  theme(legend.key.height =  unit(0.5, 'cm')) +  
  guides(color = guide_legend(order = 1),  
         shape = guide_legend(order = 2)) +     
  # add counties====     
  geom_polygon(data = counties,   
             aes(x = long, y = lat,     
                 group = group),   
             color = "gray80",  
             fill  = "NA") +   
  # add Reservation counties====  
  geom_polygon(data = rez,   
             aes(x = long, y = lat,    
                 group = group),  
             fill  = "gray90",  
             alpha = 0.5) +  
  # add Pine Ridge boundary====  
  geom_polygon(data = ridge_bound,   
             aes(x     = long,  
                 y     = lat,    
                 group = group),  
             color     = "gray50",  
             fill     = "NA") +  
  # add Rosebud boundary====  
  geom_polygon(data = rosebud_bound,   
             aes(x     = long,  
                 y     = lat,    
                 group = group),  
             color     = "gray50",  
             fill     = "NA") +  
  # add NE stateline boundary====    
  geom_line(data      = ne_bound,   
          aes(x     = long,    
              y     = lat,     
              group = group),  
          color     = "gray60") +   
  # add WY stateline boundary====    
  geom_line(data = wy_bound,   
          aes(x     = long,  
              y     = lat,    
              group = group),  
          color     = "gray60") +    
# add Pine Ridge reservation label    
  annotate("rect",  
           xmin = -102.4,  
           xmax = -101.65,  
           ymin =   43.2,  
           ymax =   43.5,   
           color = "NA",  
           fill  = "grey95", 
           alpha = 0.9) +  
  annotate("text",  
           x = -101.97, y = 43.38,  
           size = 1.7,  
           color = "gray45",  
           label = "Pine Ridge \n Reservation") +  
# add Rosebud reservation label    
  annotate("rect",  
           xmin = -100.79,  
           xmax = -100.25,  
           ymin =   43.2,  
           ymax =   43.5,   
           color = "NA",  
           fill  = "grey95", 
           alpha = 0.9) +  
  annotate("text",  
           x = -100.51, y = 43.38,  
           size = 1.7,  
           color = "gray45",  
           label = "Rosebud \n Reservation") +  
  # add stream gaging stations====     
  geom_point(data      = gages,  
           aes(dec_lon, dec_lat,   
               shape = factor(type),  
               color = ecoreg),   
           size   = 2,    
           stroke = 0.8)  

# export map & clean up   
ggplot2::ggsave(filename = "figure/gage_map.png",    
                width    = 7.5,  
                height   = 4.5,  
                units    = "in")     

# clean up global environment 
rm(bbox,   
   counties,   
   states,   
   wsd_summary,  
   ne_bound,  
   wy_bound, 
   usmap,  
   rez, 
   ridge_bound, 
   rosebud_bound)    
 
```
