==> 13-chap2a-spi.Rmd <==
---
title: "02b-chap2b-hydrology-spi"
output: pdf_document
---


```{r precip-EDA, include=FALSE, eval=FALSE}
# Purpose: EDA of precipitation data.
# Outcome: Differences among stations 1971-2018 are small.
#          less than +/-5 mm on average.

# the anomolously wet month series is dominated by May & June events.
# what drives precip during this time?  
#   In May & June the area recieves low-level moisture from the Gulf 
#   of Mexico, strong cold fronts, and active upper-level pattern 
#   leading to greater chance for convection.

# drop rows with missing data - this organizes the stations to same 
# time series length
sta_prep <- sta_raw %>%  
  drop_na %>% 
  gather(key = "sta", value = "depth", 
         -date, -year, -month) 

# create abreviated column for plotting & join
abbrev <- data.frame("sta" = c("cot", "int", "oel", "ora", "rap"),
                     "sta_abrev" = c("C", "I", "E", "O", "R")) 

sta_prep <- full_join(sta_prep, abbrev, by = "sta")

# create a year_month column
sta_prep <- sta_prep %>% 
  mutate(yr_mon = paste(year, month, sep = '_')) 

# find mean values by month & join
sta_mean <- sta_prep %>% 
  group_by(month, year) %>% 
  summarize(mean_depth = mean(depth)) %>% 
  ungroup() %>% 
  mutate(mean_depth = round(mean_depth, digits = 2)) %>% 
  mutate(yr_mon = paste(year, month, sep = '_')) %>% 
  select(-year, -month)
 
sta_prep <- full_join(sta_prep, sta_mean, by = "yr_mon") 
rm(abbrev, sta_mean)

# create a column of deviation from the mean by year & month
sta_prep <- sta_prep %>%
  mutate(deviation = depth - mean_depth) 

# plot the precip data as an overall boxplot 
ggplot(sta_prep, aes(as.factor(sta), depth)) + 
  geom_violin() + 
  geom_boxplot() + 
  scale_y_sqrt() + 
  theme_bw() + 
  ggtitle("Weather stations near Pine Ridge Reservation, SD",  
          subtitle = "1971-2018") + 
  xlab("") +  
  ylab("Monthly depth in mm") + 
  NULL 

ggplot2::ggsave(filename = "prcp_boxplot.png", 
                width = 6, height = 6, units = "in") 

# plot the precip depth as a monthly boxplot
ggplot(sta_prep, aes(sta_abrev, depth)) + 
  geom_boxplot() +
#  geom_histogram(binwidth = 1) +
  facet_grid(cols = vars(month))

# plot the precip deviations as a monthly boxplot
ggplot(sta_prep, aes(sta_abrev, deviation)) + 
  geom_boxplot() +
#  geom_histogram(binwidth = 1) +
  facet_grid(cols = vars(month))

# prepare summary of station data & clean up
sta_sum <- sta_prep %>% 
  group_by(sta, month) %>% 
  summarise(mean_depth = mean(depth),
            sd_depth = sd(depth),
            mean_dev = mean(deviation)) 

rm(sta_prep)
```

```{r setup, include=FALSE, warning=FALSE}   
knitr::opts_chunk$set(echo = TRUE)       
options(tibble.print_max = 70) # sets tibble output for printing   

# suppressMessages(require("package")) # check into this - should fix dplyr 

# Set up the package library------------------------------------ 
library("here")       # identifies where to save work  
library("rio")        # more robust I/O - to import and clean data 
library("janitor")    # tools to clean dirty data 
library("broom")      # tidies up objects 
library("corrr")      # exploring correlations
library("SCI")        # calculates SPI & RDI 
library("forecast")   # using the BoxCox function
library("mclust")     # clustering algorithms
library("factoextra") # creates exploratory plots of clusters 
library("corrplot")   # creates correlations plots
library("lubridate")  # fixes dates 
library("tidyverse")  # data munging, and tidying tools
library("magrittr")   # easier ways to work with lists
# color packages
library(colorspace) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
why_to_write <- function() 
{today <- today(tzone = "")
defense <- ymd("2019-05-22") 
until <- defense - today
print(paste("You have", until, "days until your PhD defense"
            ))
}
why_to_write()

# thoughts & ideas ---- 
#library("clValid")

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# for spatial analysis----------------------------------------------- 
#library("biogeo")     #  Functions for error detection & correction 
                      # in point-data datasets. 
                      # Includes functions for parsing & converting 
                      # coords into decimal degrees from various formats.


# check into before removing
#library("ggpubr") 
# library("munsell") # https://github.com/cwickham/munsell
# library("grateful") - not yet ready for R 3.5.0
# library("standardize")
# lmomco <- citation("lmomco")
# toBibtex(lmomco) 
# library(pdftools)
# probably delete from final version

#library("DataExplorer")
#library("lintr")
#library("test_that")
#library("jsonlite") # Convert between JSON data and R objects
#library("curl") # Drop-in replacement for base url
#library("listviewer") # htmlwidget for interactive views of R lists

# Session Info
a_session <- devtools::session_info() 

# variables----
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# a_session    list variable of session information 
# sta          precipitation station  
# spi          Standardized Precipitation Index vals  
# gage         streamflow gage 

# general modifiers:
# _meta        metadata  
# _raw         original imported data in wide format
# _gath        wide data changed to long data format
# _na          exploratory na values 
# _outlier     split outlier data 
# _typical     split not outlier data 
# _replace     intermediate value used to replace outlier value 

# _9398        first streamflow cluster by waterYear

# intermediate variables 
# abbrev       abbreviation; a look-up table 
# _mean        intermediate variable; mean value
# _sum         summary data 
# _prep        intermediate variable 

# regression variables:
# _lm          linear model 
# _aug         Added prediction info about each dataset observation 
# _fit         Fitted regression model information 
# _tidy        Summarizes information about model components 

# precipitation station identifiers:
# _cot        Cottonwood precipitation station
# _oel        Oelrichs precipitation station 
# _rap        Rapid City precipitation station 
# _int        Interior precipitation station 
# _ora        Oral precipitation station

# SCI input & intermediate variables:
# time_scale   sets the length of the averaging period 
# distrib      sets the distribution type 
# p_zero       sets a function to reduce zero-precip bias
# scale        scales input by subtract mean & divide by sd
# warn_me      sets explicit warning
# first_mon    sets the first month for SCI; based on the data

# .cot        Cottonwood precipitation station
# .oel        Oelrichs precipitation station 
# .rap        Rapid City precipitation station 
# .int        Interior precipitation station 
# .ora        Oral precipitation station
# .l          a variable as a list 
# _M          Correlation matrix 
# _check      Temporary variable used to check data

# SCI output variables
# zero_prob   probability of a zero depth month
# _att        SPI function attributes: scale, location, shape 
# _corr       final correlation dataframe
# _index      SPI index

# Gage PCA input & intermediate variables:
# _depth      streamflow depth; depth = Q/A 
# _lambda     BoxCox lambda values defined by a ML estimator 
#   _q1       one-day average streamflow volume or depth 
#   _q7       seven-day average streamflow volume or depth 
#   _q30      thirty-day average streamflow volume or depth 

# pca variables
# _input      matrix with only observations
# _meta       matrix with info about observations 
# _matrix     result of prcomp analysis as prcomp object 
# _eigen      tidy eigenvecter summary 
# _vars       tidy summary about pc loadings on variables 
# _aug        added prediction info about each dataset observation 
# _sum        summary of results


# name check 
# gage            - complete daily flow data [1990-2018] 
# gage_clust      - station groups & PCA_means 
# gage_incomp     - daily flows from partial years  
# gage_mod        - daily flows + PCA + cluster group 

# vars to add together
# gage_group        - by-group est daily flows for missing-year data 
# gage_group_part   - by-group est daily flows for partial-year data 
# gage_incomp_run02 - measured daily flows for partial year data

# temporary variables:
# gage_meta       - complete set of metadata for 26 stations ??remove?
# gage_part_miss  - station groups + days of record 
# gage_summary    - stations, water_year, days record to assist 
```

```{r spi-with-SCI, message=FALSE}   
# The following code calculates SPI using the Standardized    
# Climate Index (SCI) package.    
#   SCI is the Standardized Precipitation Index 
#   SRI is the Standardized Runoff Index (SRI) - used below 
# Another SCI package index is the Standardized Precipitation 
#   Evapotranspiration Index (SPEI) - not used

# import precipitation data---- 
sta_meta    <- as.tibble(import("data/sta_meta.csv")) %>%  
  mutate(min_date = ymd(mindate)) %>% 
  mutate(max_date = ymd(maxdate)) %>% 
  select(-mindate, -maxdate) %>% 
  mutate(dur_year = max_date - min_date) %>% 
  mutate(dur_year = days(dur_year)) %>% 
  mutate(dur_year = time_length(dur_year, unit = "year")) 
  
sta_raw     <- as.tibble(import("data/stations_monthly.csv")) %>% 
  mutate(date = ymd(date)) %>%  
  arrange(date) 
 
# checks for NA based on length of dataframe
sta_chk <- sta_raw %>% 
  gather(key = "sta", value = "depth", 
         -date, -year, -month) %>% 
  drop_na(depth)  %>% 
  mutate(sta = as.factor(sta)) %>% 
    spread(sta, depth) 

rm(sta_chk) 

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# Initial values for SCI calculations 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
time_scale <- 1  # sets the length of the averaging period 
distrib <- "pe3" # sets the distribution type 
p_zero <- TRUE   # sets a function to reduce zero-precip bias
p_zero_cm <- TRUE # ?????
scale <- "sd"    # scales input by subtract mean & divide by sd
warn_me <- TRUE  # sets explicit warning

# Set first month for each station 
first_mon_cot <- 6 
first_mon_int <- 11 
first_mon_oel <- 6 
first_mon_ora <- 5  
first_mon_rap <- 5 

# prepare station for analysis 
sta_cot <- sta_raw %>% arrange(date) %>% select(date, cot) 
sta_int <- sta_raw %>% arrange(date) %>% select(date, int) 
sta_ora <- sta_raw %>% arrange(date) %>% select(date, ora) 
sta_oel <- sta_raw %>% arrange(date) %>% select(date, oel) 
sta_rap <- sta_raw %>% arrange(date) %>% select(date, rap) 

 # change tibble to a vector as double
cot <- as.double(sta_cot$cot)
int <- as.double(sta_int$int) 
ora <- as.double(sta_ora$ora)
oel <- as.double(sta_oel$oel) 
rap <- as.double(sta_rap$rap) 

# notes: 
# A note on variable naming convention; the SCI package doesn't like 
#   snake_case variables 
# 1 month spi - Int did not close on february

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 1mon-spi-with-SCI-1mo----
# Cottonwoods SPI
#~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value)

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - throws a month 7 NA error.
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results 
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI
#~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   1mon- combine & gather the SPI variables---- 
spi_gath01 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_01 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 2mon-spi-with-SCI-2mo----

# set SCI state variables 
time_scale <- 2  # sets the length of the averaging period 

# Cottonwoods SPI - 2 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value)

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

spi.int <- transformSCI(int, first.mon = first_mon_int, obj = spi.int.l) 

spi_int <- as.tibble(spi.int) # change the vector into a tibble

spi_int <- spi_int %>% 
  rename(spi_int = value) # prepare for a later join

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value 

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value)

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   2mon- combine & gather the SPI variables---- 
spi_gath02 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_02 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 3mon-spi-with-SCI---- 
time_scale <- 3  # sets the length of the averaging period 

# Cottonwoods SPI - 3
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value)

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib,  
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value 

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   3mon - combine & gather the SPI variables---- 
spi_gath03 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_03 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 4mon-spi-with-SCI-4mo----
# notes:  ORAL HAS ONE SPI VAL OF -8
time_scale <- 4  # sets the length of the averaging period 

# Cottonwoods SPI - 4
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value)

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# prepare Interior precip for SPI transformation

# calculate SPI as a list & extract results
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value


# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results

spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value


#   4mon - combine & gather the SPI variables---- 
spi_gath04 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_04 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 

#   5mon-spi-with-SCI-5mo----
# notes: ora has an extreme low value & really high skew 
time_scale <- 5  # sets the length of the averaging period 

# Cottonwoods SPI - 5 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value)

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value



#   5mon - combine & gather the SPI variables---- 
spi_gath05 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_05 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 6mon-spi-with-SCI-6mo---- 
 time_scale <- 6  # sets the length of the averaging period 

# Cottonwoods SPI - 6
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value


#   6mon - combine & gather the SPI variables---- 
spi_gath06 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_06 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

#   9mon-spi-with-SCI-9mo----
# Notes: ora has an extreme low value & really high skew ??

# set SCI state variables 
time_scale <- 9  # sets the length of the averaging period 

# Cottonwoods SPI - 9 
#~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   9mon - combine & gather the SPI variables---- 
spi_gath09 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_09 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 12mon- spi-with-SCI-12mo----
time_scale <- 12  # sets the length of the averaging period 

# Cottonwoods SPI - 12
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)  

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results  
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value 

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value


#   12mon - combine & gather the SPI variables---- 
spi_gath12 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_12 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 18mon-spi-with-SCI-18mo----
time_scale <- 18  # sets the length of the averaging period 

# Cottonwoods SPI - 18
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

spi.int <- transformSCI(int, first.mon = first_mon_int, obj = spi.int.l) 

spi_int <- as.tibble(spi.int) # change the vector into a tibble

spi_int <- spi_int %>% 
  rename(spi_int = value) # prepare for a later join

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value 

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   18mon - combine & gather the SPI variables---- 
spi_gath18 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_18 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 24mon- spi-with-SCI-24mo---- 
time_scale <- 24  # sets the length of the averaging period 

# Cottonwoods SPI - 24
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   24mon - combine & gather the SPI variables---- 
spi_gath24 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_24 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 30mon- spi-with-SCI-30mo---- 
time_scale <- 30  # sets the length of the averaging period 

# Cottonwoods SPI - 30 
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value


#   30mon - combine & gather the SPI variables---- 
spi_gath30 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_30 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 36mon- spi-with-SCI-36mo---- 
time_scale <- 36  # sets the length of the averaging period 

# Cottonwoods SPI - 36 
#~~~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   36mon - combine & gather the SPI variables---- 
spi_gath36 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_36 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 42mon- spi-with-SCI-42mo---- 
time_scale <- 42  # sets the length of the averaging period 

# Cottonwoods SPI - 42
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   42mon - combine & gather the SPI variables---- 
spi_gath42 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_42 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 48mon- spi-with-SCI-48mo---- 
time_scale <- 48  # sets the length of the averaging period 

# Cottonwoods SPI -  
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value 


# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   48mon - combine & gather the SPI variables---- 
spi_gath48 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_48 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


# 54mon- spi-with-SCI-54mo---- 
time_scale <- 54  # sets the length of the averaging period 

# Cottonwoods SPI  
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value 

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   54mon - combine & gather the SPI variables---- 
spi_gath54 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_54 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 

# 60mon- spi-with-SCI-60mo---- 
time_scale <- 60  # sets the length of the averaging period 

# Cottonwoods SPI  
#~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 

spi.cot.l <- fitSCI(cot, first.mon = first_mon_cot , distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me) 

# create a matrix with SPI values - used below
spi.cot <- transformSCI(cot, first.mon = first_mon_cot, 
                        obj = spi.cot.l) 

# change to tibble, prepare for join
spi_cot <- as.tibble(spi.cot) %>% 
  rename(spi_cot = value) 

spi_cot <- bind_cols(sta_cot, spi_cot) # bind date, depth, SPI value

# Interior SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results 
spi.int.l <- fitSCI(int, first.mon = first_mon_int, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.int <- transformSCI(int, first.mon = first_mon_int, 
                        obj = spi.int.l) 

# change to tibble, prepare for join
spi_int <- as.tibble(spi.int) %>% 
  rename(spi_int = value)

spi_int <- bind_cols(sta_int, spi_int) # bind date, depth, SPI value

# Oelrichs SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# calculate SPI as a list & extract results
spi.oel.l <- fitSCI(oel, first.mon = first_mon_oel, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.oel <- transformSCI(oel, first.mon = first_mon_oel, 
                        obj = spi.oel.l) 

# change to tibble, prepare for join
spi_oel <- as.tibble(spi.oel) %>% 
  rename(spi_oel = value) 

spi_oel <- bind_cols(sta_oel, spi_oel) # bind date, depth, SPI value 

# Oral SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.ora.l <- fitSCI(ora, first.mon = first_mon_ora, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.ora <- transformSCI(ora, first.mon = first_mon_ora, 
                        obj = spi.ora.l) 

# change to tibble, prepare for join
spi_ora <- as.tibble(spi.ora) %>% 
  rename(spi_ora = value) 

spi_ora <- bind_cols(sta_ora, spi_ora) # bind date, depth, SPI value

# Rapid SPI 
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# calculate SPI as a list & extract results
spi.rap.l <- fitSCI(rap, first.mon = first_mon_rap, distr = distrib, 
                   time.scale = time_scale, p0 = p_zero, 
                   p0.center.mass = p_zero_cm, scaling = scale, 
                   warn	= warn_me)

# create a matrix with SPI values - used below
spi.rap <- transformSCI(rap, first.mon = first_mon_rap, 
                        obj = spi.rap.l) 

# change to tibble, prepare for join
spi_rap <- as.tibble(spi.rap) %>% 
  rename(spi_rap = value) 

spi_rap <- bind_cols(sta_rap, spi_rap) # bind date, depth, SPI value

#   60mon - combine & gather the SPI variables---- 
spi_gath60 <- list(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) %>% 
  reduce(left_join, by = "date") %>% 
  select(date, starts_with("spi")) %>% # selects the spi vals
  gather(key = sta, value = spi_index, -date) %>% 
  mutate(sta = str_remove_all(.$sta, "spi_")) %>% 
  drop_na() 

#   create a list of lists
spi_list_60 <- list(spi.cot.l, spi.int.l, spi.oel.l, 
             spi.ora.l, spi.rap.l) %>% 
  flatten() 

# remove the spi vectors & individual dataframes
rm(spi.cot, spi.int, spi.oel, spi.ora, spi.rap)
rm(spi_cot, spi_int, spi_oel, spi_ora, spi_rap) 
rm(spi.cot.l, spi.int.l, spi.oel.l, spi.ora.l, spi.rap.l) 
```

```{r join_spi_values} 

# clean up some of the unneeded variables 
rm(distrib, p_zero, p_zero_cm, scale, time_scale, warn_me, 
   sta_cot, sta_int, sta_oel, sta_ora, sta_rap, 
   cot, int, oel, ora, rap, 
   first_mon_cot, first_mon_int, first_mon_oel, 
   first_mon_ora, first_mon_rap) 

# Fix month 7 Interior 1-mo-SCI (this is actually Feb!)
# finds dates of record - note that uses Mar rather than Feb
spi_ml_fail <- spi_gath01 %>% 
  mutate(year = year(date)) %>%              # creates ymd
  mutate(mon = month(date)) %>% 
  mutate(day = day(date)) %>% 
  filter(mon == 3) %>%                       # filters dates for 'int'
  filter(sta == "int") %>% 
  mutate(mon = 2) %>% 
  mutate(date = str_c(year, mon, day, sep = "-")) %>% 
  mutate(date = ymd(date)) %>%               # creates updated months
  select(date, sta) %>% 
  group_by(sta) %>% 
  summarise(min_date = min(date))            # summarizes to min month
  
# calculates the average value for each of the dates
spi_na_fix <- spi_gath01 %>% 
  mutate(mon = month(date)) %>% 
  filter(mon == 2) %>% 
  spread(key = sta, value = spi_index) %>% 
  select(-mon) %>% 
  select(date, everything()) %>% 
  mutate(spi_index = rowMeans(.[,-1], na.rm = TRUE)) %>% 
  mutate(sta = "int") 

# filters by the min date for 'int' & prepares to append to spi_gath01
spi_na_fix <- spi_na_fix %>% 
  filter(date >= spi_ml_fail$min_date) %>% 
  select(date, sta, spi_index)

# append the missing values & clean up
spi_gath01 <- bind_rows(spi_gath01, spi_na_fix) 
rm(spi_ml_fail, spi_na_fix)

#create a list, join, rename, and gather the SPI_indices 
spi_index <- list(spi_gath01, spi_gath02, spi_gath03, spi_gath04, 
                   spi_gath05, spi_gath06, spi_gath09, spi_gath12, 
                   spi_gath18, spi_gath24, spi_gath30, spi_gath36, 
                   spi_gath42, spi_gath48, spi_gath54, spi_gath60) %>%
  reduce(left_join, by = c("date", "sta")) 

# renames as spi_length variables 
names(spi_index)[3:18] =
  c("1","2", "3", "4", "5", "6", 
    "9", "12", "18", "24", "30", "36", 
                   "42", "48","54", "60") 

# gather the dataframe 
spi_index <- spi_index %>% 
  gather(key = spi_length, val = spi_index, -date, -sta) %>% 
  mutate(spi_length = as.integer(spi_length)) 

# remove the joined dataframes
rm(spi_gath01, spi_gath02, spi_gath03, spi_gath04, 
                   spi_gath05, spi_gath06, spi_gath09, spi_gath12, 
                   spi_gath18, spi_gath24, spi_gath30, spi_gath36, 
                   spi_gath42, spi_gath48, spi_gath54, spi_gath60) 

# check for na values - these are caused by rolling averages
spi_index_na <- spi_index %>% 
  filter(is.na(spi_index) | 
           is.na(sta)) %>% 
  group_by(sta, spi_length) %>% 
  summarise(count = n()) %>% 
  ungroup() %>% 
  mutate(count_dif = spi_length - count) %>% 
  distinct(count_dif)                     # if ans = 1 then ok

# remove the na values
spi_index <- spi_index %>% 
  na.omit()

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# create a dataframe of spi attributes
spi_att01 <- pluck(spi_list_01, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 1)

spi_att02 <- pluck(spi_list_02, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 2) 

spi_att03 <- pluck(spi_list_03, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 3) 

spi_att04 <- pluck(spi_list_04, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 4) 

spi_att05 <- pluck(spi_list_05, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 5) 

spi_att06 <- pluck(spi_list_06, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 6) 

spi_att09 <- pluck(spi_list_09, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 9) 

spi_att12 <- pluck(spi_list_12, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 12) 

spi_att18 <- pluck(spi_list_18, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 18) 

spi_att24 <- pluck(spi_list_24, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 24) 

spi_att30 <- pluck(spi_list_30, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 30) 

spi_att36 <- pluck(spi_list_36, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 36) 

spi_att42 <- pluck(spi_list_42, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 42) 

spi_att48 <- pluck(spi_list_48, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 48) 

spi_att54 <- pluck(spi_list_54, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 54) 

spi_att60 <- pluck(spi_list_60, 1) %>% 
  as_tibble(rownames = "property") %>% 
  gather(key = month, value = val, -property) %>% 
  mutate(month = str_remove(.$month, "M")) %>% 
  mutate(spi_length = 60) 

#create a list, and join SPI attributes 
spi_att <- list(spi_att01, spi_att02, spi_att03, spi_att04, 
                   spi_att05, spi_att06, spi_att09, spi_att12, 
                   spi_att18, spi_att24, spi_att30, spi_att36, 
                   spi_att42, spi_att48, spi_att54, spi_att60) %>%
  reduce(left_join, by = c("property", "month")) %>%
  mutate(month = as.integer(month)) %>% 
  gather(key, spi_len, -month, -property, -starts_with("val")) %>% 
  select(-key) %>% 
  gather(key, val, -month, -property, -spi_len) %>% 
  select(-key) 

# clean up environment 
rm(spi_att01, spi_att02, spi_att03, spi_att04, 
                   spi_att05, spi_att06, spi_att09, spi_att12, 
                   spi_att18, spi_att24, spi_att30, spi_att36, 
                   spi_att42, spi_att48, spi_att54, spi_att60) 

rm(spi_list_01, spi_list_02, spi_list_03, spi_list_04, 
                   spi_list_05, spi_list_06, spi_list_09, spi_list_12, 
                   spi_list_18, spi_list_24, spi_list_30, spi_list_36, 
                   spi_list_42, spi_list_48, spi_list_54, spi_list_60) 
rm(spi_index_na)
```

```{r examine-SCI-values, eval=FALSE}  
# This code chunk examines & fixes the really large negative values 
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
# Visualize initial SPI values----
ggplot(spi_index, aes(as.integer(spi_length), spi_index)) + 
  geom_jitter() +  
  facet_grid(rows = vars(sta)) + 
  theme_bw() +
  ggtitle("Initial SPI Plot", 
          subtitle = "Typical values are between -4 to 4") 

# save intial plot
#ggplot2::ggsave(filename = "figure/initial_spi.png", 
#                width = 6, height = 6, units = "in")

# split outliers from typical values----
spi_index_outlier <- spi_index %>% 
  filter(spi_index < -4) %>% 
  as.tibble() 

spi_index_typical <- spi_index %>% 
  filter(spi_index > -4) %>%
  arrange(spi_index)

# plot outlier years----
sta_outlier <- sta_raw %>% 
  gather(key = sta, value = depth, -date, -year, -month) %>% 
  filter(date == "1977-02-01" | 
    date == "2005-01-01" | 
    date == "2007-01-01" | 
    date == "2012-11-01" | 
    date == "1976-05-01" |
    date == "1981-09-01" |
    date == "2007-02-01"
      ) 

ggplot(sta_outlier, aes(date, depth, color = factor(sta))) + 
  geom_point() + 
  theme_bw() +
  ggtitle("Plot of outlier years") 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# spi_index_typical - calculated above 
# these are the min SPI vals from Cottonwood during depression 
#  date       sta   spi_index spi_length 
#   <date>     <chr>     <dbl>      <dbl> 
# 1 1936-10-01 cot       -3.31          6 
# 2 1937-01-01 cot       -3.27          9 
# 3 1936-09-01 cot       -3.26          4 
# 4 1936-09-01 cot       -3.25          5 
# 5 1936-10-01 cot       -3.25          9  

# spi_index_outlier - calculated above 
# date       sta   spi_index spi_length
#  <date>     <chr>     <dbl>      <dbl>
#1 1977-02-01 ora       -8.03          5   * caused by a zero depth
#2 2005-01-01 ora       -7.51         12
#3 2007-01-01 ora       -8.01          4
#4 2012-09-01 int       -8.03          1 
#5 2012-11-01 ora       -7.71         12 
#6 1976-05-01 ora       -7.93         30
#7 1981-09-01 ora       -7.73         36
#8 2007-02-01 ora       -8.01         36

# Outlier values could be approximated by the mean SPI  
# or the nearest neighbor (Oral <- Oelrichs) - I used mean SPI  
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 

# create replacement values
spi_index_replace <- spi_index_typical %>%          # pulls outliers
  filter(date == "1977-02-01" &  spi_length == 5   | 
           date == "2005-01-01" & spi_length == 12 |  
           date == "2007-01-01" & spi_length == 4  | 
           date == "2012-11-01" & spi_length == 12 | 
           date == "2012-09-01" & spi_length == 1  | 
           date == "1976-05-01" & spi_length == 30 | 
           date == "1981-09-01" & spi_length == 30 | 
           date == "2007-02-01" & spi_length == 30  
         )  %>% 
    group_by(date, spi_length) %>% 
  summarize(spi_index = mean(spi_index)) %>%
  ungroup()                      # calculates mean values 

# update replacement values & join with typical values 
spi_index_replace <- spi_index_outlier %>% 
  select(date, sta) %>% 
  full_join(., spi_index_replace, by = "date")

spi_index2 <- bind_rows(spi_index_typical, spi_index_replace) %>% 
  arrange(sta, spi_length, date)

# check the values after the fix - the outlier values are now fixed 
# spi_index2
# A tibble: 48,011 x 4
#   date       sta   spi_index spi_length
#   <date>     <chr>     <dbl>      <dbl>
# 1 1936-10-01 cot       -3.31          6
# 2 1937-01-01 cot       -3.27          9
# 3 1936-09-01 cot       -3.26          4
# 4 1936-09-01 cot       -3.25          5
# 5 1936-10-01 cot       -3.25          9

# update the index in memory 
spi_index <- spi_index2 %>% 
  arrange(spi_index)

# clean up
rm(spi_index_outlier, spi_index_replace, 
   spi_index_typical, spi_index2, sta_outlier)
```

```{r final-spi-plot}  
spi_index_plot <- spi_index %>% 
  filter(date > "1990-01-01")  %>% 
  filter(spi_length == 1 | 
           spi_length == 6 |
           spi_length == 12 | 
           spi_length == 24 |
           spi_length == 36 |
           spi_length == 48 | 
           spi_length == 60
         ) %>%
  mutate(year = year(date))

ggplot(spi_index_plot, aes(year, spi_index)) + 
  geom_point() + 
  geom_smooth() + 
  geom_hline(yintercept = 0, color = "gray") +
  facet_grid(rows = vars(spi_length), 
             cols = vars(sta)) +
  theme_classic() + 
  ggtitle("SPI Index values for Pine Ridge reservation stations",
          subtitle = "1990-01-01 to 2018-05-30")

ggplot2::ggsave(filename = "figure/spi_vals.png", 
                width = 8, height = 8, units = "in") 

# the model shows spi values are highly correlated with only real 
# difference being the number of years in the calculation.
# However, individual droughts differ in space & time 

#ggplot(spi_index_plot, aes(date, spi_index)) + 
#  geom_line() + 
#  facet_wrap(vars(sta)) + 
#  theme_classic() +
#  geom_hline(yintercept = 0) + 
#  scale_x_date(limits = as.Date(c('1980-01-01','2018-01-01')), 
#                 date_breaks = "1 year")  
#labels = date_format("%b-%Y")
```

```{r spi-correlation-analysis}
# visually check results
#spi_index_plot <- spi_index %>% 
#  filter(spi_length == 12)
  
#ggplot(spi_index_plot, aes(date, spi_index)) + 
#  geom_line() +
#  facet_wrap(vars(sta)) + 
#  theme_classic() +
#  geom_hline(yintercept = 0, aes)

#ggplot2::ggsave(path = "figure/", filename = "spi_1mo.png", 
#                width = 6, height = 6, units = "in")  

# create correlation matrix inputs
spi_M <- spi_index %>% 
  select(-date) %>% 
  group_by(sta, spi_length) %>% 
  mutate(grouped_id = row_number()) %>% 
  spread(key = sta, value = spi_index) %>% 
  drop_na() %>% 
  ungroup() %>% 
  select(-grouped_id) %>% 
  select(spi_length, cot, int, oel, ora, rap) # ensure vars order 

# create correlation matrix names from the correlation matrix vars
spi_M_names <- spi_M %>% 
  filter(spi_length == 1) %>% 
  cor() %>% 
  as.tibble() %>% 
  names() %>% 
  as.tibble() %>% 
  slice(-1) %>% 
  mutate(value2 = value) %>% 
  mutate(value3 = value) %>%  
  mutate(value4 = value) %>%
  mutate(value5 = value) %>% 
  gather(key, sta2) %>% 
  select(-key) %>% 
  rownames_to_column() 

# create second station names column 
spi_M_names2 <- spi_M_names %>% 
  arrange(sta2) %>% 
  rename(sta1 = sta2) %>% 
  select(-rowname)

# bind the names columns
spi_M_names <- bind_cols(spi_M_names, spi_M_names2)  
rm(spi_M_names2)

# create a correlation matrix from SPI vals
spi_M <- spi_M %>% 
  split(.$spi_length) %>% 
  purrr::map_dfr(~ cor(.)) %>% 
  drop_na() %>% 
  slice(-1) %>% 
  rownames_to_column() 

# bind names to the correlation matrix 
spi_M <- full_join(spi_M_names, spi_M, by = "rowname")
spi_M <- spi_M %>% 
  select(sta1, sta2, everything()) %>%
  select(-rowname)
  
# prepare lookup table of lat-lons
sta_loc <- sta_meta %>% 
  arrange(name) %>%
  mutate(sta = c("cot", "int", "oel", "ora", "rap")) %>% 
  select(sta, lat, lon, dur_year)

# join the 'from' lat lons
spi_corr <- full_join(spi_M, sta_loc, by = c("sta1" = "sta")) %>% 
  rename(lat1 = lat) %>% 
  rename(lon1 = lon) 

# join the 'to' lat lons
spi_corr <- full_join(spi_corr, sta_loc, by = c("sta2" = "sta")) %>% 
  rename(lat2 = lat) %>% 
  rename(lon2 = lon) %>% 
  mutate(year_dif = abs(dur_year.x - dur_year.y))
  

# convert the 'to' and 'from' lat-lons to northings & eastings & distance
lat_to_km <- 111.03 # 1 degree lat to km @ lat 40-degrees 
lon_to_km <- 85.39  # 1 degree lon to km @ lat 40-degrees 
  
spi_corr <- spi_corr %>% 
  mutate(northing = abs((lat1 - lat2)) * lat_to_km) %>% 
  mutate(easting = abs((lon1 - lon2)) * lat_to_km) %>% 
  mutate(distance = sqrt(northing^2 + easting^2)) %>% 
  select(year_dif, everything()) %>%
  select(-(lat1:easting)) %>% 
  mutate(stations = paste(sta1, sta2, sep = "_")) %>% 
  gather(key = spi_length, value = pears_r, -distance, 
         -stations, -sta1, -sta2, -year_dif) %>% 
  filter(distance > 0) %>% 
  mutate(spi_length = as.double(spi_length))


# clean up the global environment 
rm(spi_M, spi_M_names, sta_loc, lat_to_km, lon_to_km)

# model effect of averaging time & distance on correlation----
# fit a linear model
spi_lm <- lm(pears_r ~ distance + spi_length + year_dif, 
             data = spi_corr)

# augment & gather the original data
spi_corr_aug <- augment(spi_lm, spi_corr) 

spi_corr_gath <- spi_corr_aug %>% 
  select(-(sta1:sta2)) %>%
  select(year_dif:.fitted) %>% 
  gather(key = factor, val, -stations, -pears_r) %>% 
  mutate(val = as.double(val))  

# plot the original data and fitted model for SPI----  
ggplot(spi_corr_gath, aes(val, pears_r)) + 
  geom_point(aes(color = factor(stations))) +
  facet_wrap(ncol = 1, vars(factor), scales = "free") +
    geom_smooth(method = lm) + 
  theme_classic()

spi_lm_fit <- glance(spi_lm) 

spi_lm_tidy <- tidy(spi_lm) %>% 
  mutate(
    low = estimate - std.error,
    high = estimate + std.error
  )

# clean-up Global Environment----
spi_corr <- spi_corr_aug 
rm(spi_corr_aug, spi_corr_gath, spi_lm ,spi_lm_fit)


rm(spi_corr, spi_lm_tidy)
```

```{r SCI-diagnostic-plots, eval=FALSE}
# Diagnostic plots of SPI transforms 

# PE3 scale plot---- 
spi_scale <- spi_att %>% 
  filter(property == "scale")

ggplot(spi_scale, aes(as.factor(month), spi_value)) +  
  geom_boxplot() +
  facet_grid(rows = vars(sta),
             cols = vars(spi_length)) + 
  theme_classic() 

ggplot2::ggsave(filename = "figure/spi_scale.png", 
                width = 6, height = 6, units = "in") 

# PE3 location plot---- 
spi_location <- spi_att %>% 
  filter(property == "location")

ggplot(spi_location, aes(as.factor(month), spi_value)) + 
  geom_boxplot() +
  facet_grid(rows = vars(sta),
             cols = vars(spi_length)) + 
  theme_classic() 

ggplot2::ggsave(filename = "figure/spi_location.png", 
                width = 6, height = 6, units = "in") 

# PE3 shape plot---- 
spi_shape <- spi_att %>% 
  filter(property == "shape")

ggplot(spi_shape, aes(as.factor(month), spi_value)) + 
  geom_boxplot() +
  facet_grid(rows = vars(sta),
             cols = vars(spi_length)) + 
  theme_classic() 

ggplot2::ggsave(filename = "figure/spi_shape.png", 
                width = 6, height = 6, units = "in") 

# clean up Global Environment----
rm(spi_location, spi_shape, spi_scale) 
```
